requirements.txt
python-telegram-bot==21.6
google-genai>=0.3.0
python-dotenv>=1.0.1
aiohttp>=3.9.5
